Move platform specific tests into platform folders
