+++ Resources/blacklist.js
@@ -7,12 +7,15 @@
 
 module.exports = {
 	'android': [
-
+		'ios_tests',
+		'windows_tests',
 	],
 	'ios':[
-
+		'android_tests',
+		'windows_tests',
 	],
 	'windows': [
-
+		'android_tests',
+		'ios_tests'
 	]
 };
