+++ Resources/ios_tests/ti.app.ios.searchquery.test.js
@@ -8,7 +8,7 @@
 /* global Ti */
 /* eslint no-unused-expressions: "off" */
 'use strict';
-var should = require('./utilities/assertions');
+var should = require('../utilities/assertions');
 
 describe.ios('Titanium.App.iOS', function () {
 	// TIMOB-23542 test searchQuery
@@ -21,9 +21,9 @@ describe.ios('Titanium.App.iOS', function () {
 describe.ios('Titanium.App.iOS.SearchQuery', function () {
 	it('constructor', function () {
 		var searchQuery = Ti.App.iOS.createSearchQuery({
-						queryString: 'title == "Titanium*"',
-						attributes: ["title", "displayName", "keywords", "contentType"]
-				});
+			queryString: 'title == "Titanium*"',
+			attributes: [ 'title', 'displayName', 'keywords', 'contentType' ]
+		});
 		should(searchQuery).have.readOnlyProperty('apiName').which.is.a.String;
 		should(searchQuery.apiName).be.eql('Ti.App.iOS.SearchQuery');
 		should(searchQuery.attributes).be.an.Array;
